# src/main/java/com/sabrina/librarymanagementsystem/controller/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.librarymanagementsystem.controller.api;


import com.sabrina.librarymanagementsystem.controller.dto.BookRequest;
import com.sabrina.librarymanagementsystem.controller.dto.BookResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-26T19:55:51.535413611+02:00[Europe/Andorra]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Books", description = "Book management")
public interface BooksApi {

    public static final String PATH_API_BOOKS_GET = "/api/books";
    /**
     * GET /api/books : Get all books
     *
     * @return List of books (status code 200)
     */
    @Operation(
        operationId = "apiBooksGet",
        summary = "Get all books",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of books", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = BookResponse.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_API_BOOKS_GET,
        produces = { "application/json" }
    )
    ResponseEntity<List<com.sabrina.librarymanagementsystem.controller.dto.BookResponse>> apiBooksGet(
        
    );


    public static final String PATH_API_BOOKS_ID_DELETE = "/api/books/{id}";
    /**
     * DELETE /api/books/{id} : Delete book
     *
     * @param id  (required)
     * @return Book deleted (status code 204)
     */
    @Operation(
        operationId = "apiBooksIdDelete",
        summary = "Delete book",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Book deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = BooksApi.PATH_API_BOOKS_ID_DELETE
    )
    
    ResponseEntity<Void> apiBooksIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_BOOKS_ID_GET = "/api/books/{id}";
    /**
     * GET /api/books/{id} : Get book by id
     *
     * @param id (required)
     * @return Book found (status code 200)
     */
    @Operation(
        operationId = "apiBooksIdGet",
        summary = "Get book by id",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Book found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BookResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_API_BOOKS_ID_GET,
        produces = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.BookResponse> apiBooksIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_BOOKS_ID_PUT = "/api/books/{id}";
    /**
     * PUT /api/books/{id} : Update book
     *
     * @param id          (required)
     * @param bookRequest (required)
     * @return Book updated (status code 200)
     */
    @Operation(
        operationId = "apiBooksIdPut",
        summary = "Update book",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Book updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BookResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = BooksApi.PATH_API_BOOKS_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.BookResponse> apiBooksIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "BookRequest", description = "", required = true) @Valid @RequestBody BookRequest bookRequest
    );


    public static final String PATH_API_BOOKS_POST = "/api/books";
    /**
     * POST /api/books : Create a book
     *
     * @param bookRequest (required)
     * @return Book created (status code 201)
     */
    @Operation(
        operationId = "apiBooksPost",
        summary = "Create a book",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Book created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BookResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_API_BOOKS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.BookResponse> apiBooksPost(
        @Parameter(name = "BookRequest", description = "", required = true) @Valid @RequestBody BookRequest bookRequest
    );

}