# src/main/java/com/sabrina/librarymanagementsystem/controller/api/CategoriesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.librarymanagementsystem.controller.api;


import com.sabrina.librarymanagementsystem.controller.dto.CategoryRequest;
import com.sabrina.librarymanagementsystem.controller.dto.CategoryResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-27T15:23:07.295255320+02:00[Europe/Andorra]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Categories", description = "Category management")
public interface CategoriesApi {

    public static final String PATH_API_CATEGORIES_GET = "/api/categories";
    /**
     * GET /api/categories : Get all categories
     *
     * @return List of categories (status code 200)
     */
    @Operation(
        operationId = "apiCategoriesGet",
        summary = "Get all categories",
        tags = { "Categories" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of categories", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CategoryResponse.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CategoriesApi.PATH_API_CATEGORIES_GET,
        produces = { "application/json" }
    )
    ResponseEntity<List<com.sabrina.librarymanagementsystem.controller.dto.CategoryResponse>> apiCategoriesGet(
        
    );


    public static final String PATH_API_CATEGORIES_ID_DELETE = "/api/categories/{id}";
    /**
     * DELETE /api/categories/{id} : Delete category
     *
     * @param id  (required)
     * @return Category deleted (status code 204)
     */
    @Operation(
        operationId = "apiCategoriesIdDelete",
        summary = "Delete category",
        tags = { "Categories" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Category deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = CategoriesApi.PATH_API_CATEGORIES_ID_DELETE
    )
    
    ResponseEntity<Void> apiCategoriesIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_CATEGORIES_ID_GET = "/api/categories/{id}";
    /**
     * GET /api/categories/{id} : Get category by id
     *
     * @param id (required)
     * @return Category found (status code 200)
     */
    @Operation(
        operationId = "apiCategoriesIdGet",
        summary = "Get category by id",
        tags = { "Categories" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Category found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CategoriesApi.PATH_API_CATEGORIES_ID_GET,
        produces = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.CategoryResponse> apiCategoriesIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_CATEGORIES_ID_PUT = "/api/categories/{id}";
    /**
     * PUT /api/categories/{id} : Update category
     *
     * @param id              (required)
     * @param categoryRequest (required)
     * @return Category updated (status code 200)
     */
    @Operation(
        operationId = "apiCategoriesIdPut",
        summary = "Update category",
        tags = { "Categories" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Category updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = CategoriesApi.PATH_API_CATEGORIES_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.CategoryResponse> apiCategoriesIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "CategoryRequest", description = "", required = true) @Valid @RequestBody CategoryRequest categoryRequest
    );


    public static final String PATH_API_CATEGORIES_POST = "/api/categories";
    /**
     * POST /api/categories : Create a category
     *
     * @param categoryRequest (required)
     * @return Category created (status code 201)
     */
    @Operation(
        operationId = "apiCategoriesPost",
        summary = "Create a category",
        tags = { "Categories" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Category created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CategoriesApi.PATH_API_CATEGORIES_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<com.sabrina.librarymanagementsystem.controller.dto.CategoryResponse> apiCategoriesPost(
        @Parameter(name = "CategoryRequest", description = "", required = true) @Valid @RequestBody CategoryRequest categoryRequest
    );

}