# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: REST API for managing a library.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /api/books:
    post:
      tags:
        - Books
      summary: Create a book
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'

      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'


    get:
      tags:
        - Books
      summary: Get all books
      security:
        - bearerAuth: [ ]

      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
  /api/books/{id}:
    get:
      tags:
        - Books
      summary: Get book by id
      security:
        - bearerAuth: [ ]

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64

      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'


    put:
      tags:
        - Books
      summary: Update book
      security:
        - bearerAuth: [ ]

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'

      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'


    delete:
      tags:
        - Books
      summary: Delete book
      security:
        - bearerAuth: [ ]

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64

      responses:
        '204':
          description: Book deleted
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'

    post:
      tags:
        - Categories
      summary: Create a category
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by id
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

    put:
      tags:
        - Categories
      summary: Update category
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

    delete:
      tags:
        - Categories
      summary: Delete category
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Category deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookRequest:
      type: object
      required:
        - title
        - isbn
        - publicationYear
        - authorId
        - categoryId

      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Clean Code

        isbn:
          type: string
          minLength: 10
          maxLength: 20
          example: 9780132350884

        publicationYear:
          type: integer
          minimum: 1900
          example: 2008

        authorId:
          type: integer
          format: int64
          minimum: 1
          example: 1

        categoryId:
          type: integer
          format: int64
          minimum: 1
          example: 2
    BookResponse:
      type: object

      properties:
        id:
          type: integer
          format: int64

        title:
          type: string

        isbn:
          type: string

        publicationYear:
          type: integer

        available:
          type: boolean

        authorName:
          type: string

        categoryName:
          type: string
...
    CategoryRequest:
      type: object
      required:
        - name

      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Programming
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        name:
          type: string
          example: Programming
...